+++ front/app/modules/commercial/representativeness/admin/hooks/parseReferenceData.ts
